ir_to_mesa: Fix warning in last commit.
I swear there was some git option for "don't push things when you've
got uncommitted changes", but I can't find it now.

+++ src/mesa/program/ir_to_mesa.cpp
@@ -892,7 +892,7 @@ ir_to_mesa_visitor::visit(ir_variable *ir)
 	       if (storage->index == -1) {
 		  storage->index = index;
 	       } else {
-		  assert(index == (storage->index +
+		  assert(index == ((int)storage->index +
 				   a * statevar->num_elements + i));
 	       }
 	    } else {
